+++ apps/avifdec.c
@@ -34,7 +34,7 @@ static void syntax(void)
     printf("    -j,--jobs J       : Number of jobs (worker threads), or 'all' to potentially use as many cores as possible. (Default: all)\n");
     printf("    -c,--codec C      : Codec to use (choose from versions list below)\n");
     printf("    -d,--depth D      : Output depth, either 8 or 16. (PNG only; For y4m, depth is retained, and JPEG is always 8bpc)\n");
-    printf("                        Output depth set to 16 enables Sample Transform decoding no matter the output format.\n");
+    printf("    --sato            : Enable Sample Transforms decoding (e.g. 16-bit AVIF)\n");
     printf("    -q,--quality Q    : Output quality in 0..100. (JPEG only, default: %d)\n", DEFAULT_JPEG_QUALITY);
     printf("    --png-compress L  : PNG compression level in 0..9 (PNG only; 0=none, 9=max). Defaults to libpng's builtin default\n");
     printf("    -u,--upsampling U : Chroma upsampling (for 420/422). One of 'automatic' (default), 'fastest', 'best', 'nearest', or 'bilinear'\n");
@@ -89,6 +89,7 @@ int main(int argc, char * argv[])
     const char * inputFilename = NULL;
     const char * outputFilename = NULL;
     int requestedDepth = 0;
+    avifBool enableSampleTransforms = AVIF_FALSE;
     int jobs = -1;
     int jpegQuality = DEFAULT_JPEG_QUALITY;
     int pngCompressionLevel = -1; // -1 is a sentinel to avifPNGWrite() to skip calling png_set_compression_level()
@@ -169,6 +170,8 @@ int main(int argc, char * argv[])
                 fprintf(stderr, "ERROR: invalid depth: %s\n", arg);
                 return 1;
             }
+        } else if (!strcmp(arg, "--sato")) {
+            enableSampleTransforms = AVIF_TRUE;
         } else if (!strcmp(arg, "-q") || !strcmp(arg, "--quality")) {
             NEXTARG();
             jpegQuality = atoi(arg);
@@ -314,7 +317,7 @@ int main(int argc, char * argv[])
     decoder->allowProgressive = allowProgressive;
     if (infoOnly) {
         decoder->imageContentToDecode |= AVIF_IMAGE_CONTENT_GAIN_MAP | AVIF_IMAGE_CONTENT_SAMPLE_TRANSFORMS;
-    } else if (requestedDepth == 16) {
+    } else if (enableSampleTransforms) {
         decoder->imageContentToDecode |= AVIF_IMAGE_CONTENT_SAMPLE_TRANSFORMS;
     }